+++ libs/block-editor/src/lib/demo/block-editor.component.html
@@ -1,3 +1,2 @@
 
-{{editor.getJSON() | json }}
 <tiptap-editor [editor]="editor" [(ngModel)]="value"></tiptap-editor>
