+++ libs/block-editor/src/lib/extensions/blocks/image-block/image-block.component.html
@@ -1,1 +1,1 @@
-<img [src]="'http://localhost:8080'+data.asset">
+<img [src]="data.asset">
