+++ libs/block-editor/src/lib/extensions/services/dot-image/dot-image.service.ts
@@ -1,24 +1,19 @@
 import { Injectable } from '@angular/core';
-import { DotAssetService, DotUploadService } from '@dotcms/utils';
+import { uploadBinaryFile } from '@dotcms/utils';
 import { from, Observable, Observer } from 'rxjs';
-import { switchMap } from 'rxjs/operators';
+import { map } from 'rxjs/operators';
 import { DotCMSTempFile } from '@dotcms/dotcms-models';
+import { HttpClient } from '@angular/common/http';
 
 @Injectable({
     providedIn: 'root'
 })
 export class DotImageService {
-    uploadService;
-    assetService;
-
-    constructor() {
-        this.uploadService = new DotUploadService();
-        this.assetService = new DotAssetService();
-    }
+    constructor(private http: HttpClient) {}
 
     get(data: File | File[], maxSize?: string): Observable<any> {
         return this.setTempResource(data, maxSize).pipe(
-            switchMap((response: DotCMSTempFile | DotCMSTempFile[]) => {
+            map((response: DotCMSTempFile | DotCMSTempFile[]) => {
                 const files = Array.isArray(response) ? response : [response];
                 return new Observable((observer: Observer<object>) => {
                     let contentlets = [];
@@ -29,23 +24,17 @@ export class DotImageService {
                             hostFolder: '',
                             indexPolicy: 'WAIT_FOR'
                         });
-
-                        fetch(
-                            'http://localhost:8080/api/v1/workflow/actions/default/fire/PUBLISH',
+                        return this.http.post(
+                            '/api/v1/workflow/actions/default/fire/PUBLISH',
+                            JSON.stringify({ contentlets }),
                             {
-                                method: 'POST',
                                 headers: {
                                     Origin: window.location.hostname,
                                     'Content-Type': 'application/json;charset=UTF-8',
                                     Authorization: 'Basic YWRtaW5AZG90Y21zLmNvbTphZG1pbg=='
-                                },
-                                body: JSON.stringify({ contentlets })
+                                }
                             }
-                        )
-                            .then((response) => response.json())
-                            .then((data) => {
-                                observer.next(data.entity.results);
-                            });
+                        );
                     });
                 });
             })
@@ -53,6 +42,6 @@ export class DotImageService {
     }
 
     private setTempResource(data: File | File[], maxSize?: string): Observable<any> {
-        return from(this.uploadService.uploadBinaryFile(data, () => {}, maxSize));
+        return from(uploadBinaryFile(data, () => {}, maxSize));
     }
 }
