+++ libs/utils/src/index.ts
@@ -1,2 +1,2 @@
-export * from './lib/services/dot-asset/dot-asset.service';
-export * from './lib/services/dot-upload/dot-upload.service';
+export * from './lib/services/dot-asset.service';
+export * from './lib/services/dot-temp-file.service';
